+++ src/GitHub/REST/Monad/Class.hs
@@ -13,7 +13,7 @@ module GitHub.REST.Monad.Class
   ( MonadGitHubREST(..)
   ) where
 
-import Control.Monad (void, (<=<))
+import Control.Monad (void)
 import Control.Monad.Trans.Class (lift)
 import Control.Monad.Trans.Except (ExceptT)
 import Control.Monad.Trans.Identity (IdentityT)
@@ -29,8 +29,6 @@ import Data.Aeson (FromJSON, Value)
 #if !MIN_VERSION_base(4,11,0)
 import Data.Monoid ((<>))
 #endif
-import Data.Text (Text)
-import qualified Data.Text as Text
 
 import GitHub.REST.Endpoint
 import GitHub.REST.PageLinks (PageLinks(..))
@@ -66,24 +64,16 @@ import GitHub.REST.PageLinks (PageLinks(..))
 -- >   , ghData = []
 -- >   }
 class Monad m => MonadGitHubREST m where
-  {-# MINIMAL queryGitHubPage' #-}
+  {-# MINIMAL queryGitHubPage #-}
 
-  -- | Query GitHub, returning @Right (payload, links)@ if successful, where @payload@ is the
+  -- | Query GitHub, returning @(payload, links)@ if successful, where @payload@ is the
   -- response that GitHub sent back and @links@ containing any pagination links GitHub may have
   -- sent back. If the response could not be decoded as JSON, returns
   -- @Left (error message, response from server)@.
   --
-  -- Errors on network connection failures or if GitHub sent back an error message. Use `githubTry`
-  -- if you wish to handle GitHub errors.
-  queryGitHubPage' :: FromJSON a => GHEndpoint -> m (Either (Text, Text) (a, PageLinks))
-
-  -- | 'queryGitHubPage'', except calls 'fail' if JSON decoding fails.
+  -- Errors on network connection failures, if GitHub sent back an error message, or if the response
+  -- could not be decoded as JSON. Use `githubTry` if you wish to handle GitHub errors.
   queryGitHubPage :: FromJSON a => GHEndpoint -> m (a, PageLinks)
-  queryGitHubPage = either fail' pure <=< queryGitHubPage'
-    where
-      fail' (message, response) =
-        let ellipses s = if Text.length s > 100 then take 100 (Text.unpack s) ++ "..." else Text.unpack s
-        in error $ "Could not decode response:\nmessage = " ++ ellipses message ++ "\nresponse = " ++ ellipses response
 
   -- | 'queryGitHubPage', except ignoring pagination links.
   queryGitHub :: FromJSON a => GHEndpoint -> m a
@@ -107,31 +97,31 @@ class Monad m => MonadGitHubREST m where
 {- Instances for common monad transformers -}
 
 instance MonadGitHubREST m => MonadGitHubREST (ReaderT r m) where
-  queryGitHubPage' = lift . queryGitHubPage'
+  queryGitHubPage = lift . queryGitHubPage
 
 instance MonadGitHubREST m => MonadGitHubREST (ExceptT e m) where
-  queryGitHubPage' = lift . queryGitHubPage'
+  queryGitHubPage = lift . queryGitHubPage
 
 instance MonadGitHubREST m => MonadGitHubREST (IdentityT m) where
-  queryGitHubPage' = lift . queryGitHubPage'
+  queryGitHubPage = lift . queryGitHubPage
 
 instance MonadGitHubREST m => MonadGitHubREST (MaybeT m) where
-  queryGitHubPage' = lift . queryGitHubPage'
+  queryGitHubPage = lift . queryGitHubPage
 
 instance (Monoid w, MonadGitHubREST m) => MonadGitHubREST (Lazy.RWST r w s m) where
-  queryGitHubPage' = lift . queryGitHubPage'
+  queryGitHubPage = lift . queryGitHubPage
 
 instance (Monoid w, MonadGitHubREST m) => MonadGitHubREST (Strict.RWST r w s m) where
-  queryGitHubPage' = lift . queryGitHubPage'
+  queryGitHubPage = lift . queryGitHubPage
 
 instance MonadGitHubREST m => MonadGitHubREST (Lazy.StateT s m) where
-  queryGitHubPage' = lift . queryGitHubPage'
+  queryGitHubPage = lift . queryGitHubPage
 
 instance MonadGitHubREST m => MonadGitHubREST (Strict.StateT s m) where
-  queryGitHubPage' = lift . queryGitHubPage'
+  queryGitHubPage = lift . queryGitHubPage
 
 instance (Monoid w, MonadGitHubREST m) => MonadGitHubREST (Lazy.WriterT w m) where
-  queryGitHubPage' = lift . queryGitHubPage'
+  queryGitHubPage = lift . queryGitHubPage
 
 instance (Monoid w, MonadGitHubREST m) => MonadGitHubREST (Strict.WriterT w m) where
-  queryGitHubPage' = lift . queryGitHubPage'
+  queryGitHubPage = lift . queryGitHubPage