+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 ## Upcoming
 
+## 1.1.0
+
+* Rename `GitHubState` to `GitHubSettings`
+* Remove `queryGitHubPage'` -- implement `queryGitHubPage` in `MonadGitHubREST` instead.
+* Expose `queryGitHubPageIO` if users want to manually implement `MonadGitHubREST`
+
 ## 1.0.3
 
 * Fix goldens after GitHub changed documentation URL